Changed unimodality pattern to a bar plot with significantly easier to understand visualization, refactored patterns.py into multiple files (instead of the 1 file with well over 1000 lines at this point).
