+++ src/external_explainers/metainsight_explainer/patterns/__init__.py
@@ -1,0 +1,5 @@
+from external_explainers.metainsight_explainer.patterns.unimodality_pattern import UnimodalityPattern
+from external_explainers.metainsight_explainer.patterns.outlier_pattern import OutlierPattern
+from external_explainers.metainsight_explainer.patterns.trend_pattern import TrendPattern
+from external_explainers.metainsight_explainer.patterns.cycle_pattern import CyclePattern
+from external_explainers.metainsight_explainer.patterns.pattern_base_classes import PatternInterface
